Replace render system with render function
