+++ script/main.js
@@ -18,6 +18,7 @@ import updateActionTicks from './eventHandlers/updateActionTicks.js';
 import updatePositionOfGameObjects from './eventHandlers/updatePositionOfGameObjects.js';
 import updatePositioningOfViewports from './library/core/eventHandlers/updatePositioningOfViewports.js';
 import animateGameObjects from './library/core/eventHandlers/animateGameObjects.js';
+import { drawFrame } from './library/core/system/RenderSystem.js';
 
 import ActionSystem from './system/ActionSystem.js';
 import AttackSystem from './system/AttackSystem.js';
@@ -27,7 +28,6 @@ import DeathSystem from './system/DeathSystem.js';
 import EquipmentDamageSystem from './system/EquipmentDamageSystem.js';
 import LogSystem from './system/LogSystem.js';
 import PlayerControlSystem from './system/PlayerControlSystem.js';
-import RenderSystem from './library/core/system/RenderSystem.js';
 
 let game = new Game(store, 'Sorcerer', document.querySelector('.canvas__sorcerer'), { scale: 4 });
 setupInterfaceEvents(game);
@@ -95,8 +95,8 @@ game.addSystem(new DeathSystem(game));
 game.addSystem(new LogSystem(game));
 game.addEventHandler('beforeDraw', updatePositionOfGameObjects);
 game.addEventHandler('beforeDraw', updatePositioningOfViewports);
-game.addSystem(new RenderSystem(game));
 game.addEventHandler('beforeDraw', animateGameObjects);
+game.addEventHandler('draw', drawFrame);
 
 game.start();
